# Optimize multi serialization