A handful of changes to optimize serialization of a multi into cellblocks with a big emphasis on reducing allocations.
The current state of the code does about one allocation per value in a call plus around 6 allocations per call. The MultiToProtoLarge benchmark creates a multi with 1000 calls, each with 5 values and would result in 11.2k allocations. 
The code in this change brings it to roughly 4 allocations per call in a multi and removes the per-value allocation, bringing the allocation count to 4.2k in MultiToProtoLarge.

The MultiToProto (not large) benchmark creates a multi with 5 calls, each with 0 or 1 value. It does not see as much benefits from these changes. The common case is sending a large multi, so I think that tradeoff is worth it.

```
name                             old time/op    new time/op    delta
MultiToProto/cellblocks-16         5.11µs ± 1%    4.94µs ± 0%   -3.39%  (p=0.000 n=10+10)
MultiToProto/no_cellblocks-16      4.67µs ± 1%    4.67µs ± 0%     ~     (p=0.589 n=9+10)
MultiToProtoLarge/cellblocks-16    1.69ms ± 1%    1.13ms ± 0%  -33.36%  (p=0.000 n=10+10)

name                             old alloc/op   new alloc/op   delta
MultiToProto/cellblocks-16         3.02kB ± 0%    2.93kB ± 0%   -3.06%  (p=0.000 n=10+10)
MultiToProto/no_cellblocks-16      3.44kB ± 0%    3.56kB ± 0%   +3.49%  (p=0.000 n=10+10)
MultiToProtoLarge/cellblocks-16    1.68MB ± 0%    0.91MB ± 0%  -45.87%  (p=0.000 n=10+10)

name                             old allocs/op  new allocs/op  delta
MultiToProto/cellblocks-16           53.0 ± 0%      43.0 ± 0%  -18.87%  (p=0.000 n=10+10)
MultiToProto/no_cellblocks-16        59.0 ± 0%      54.0 ± 0%   -8.47%  (p=0.000 n=10+10)
MultiToProtoLarge/cellblocks-16     11.2k ± 0%      4.2k ± 0%  -62.49%  (p=0.000 n=10+10)
```